Fixing list comprehension mutation and adding unit test.
The previous mutation was not happening on the first pass, because `in`
was mutated to `not in` in memory, which meant the list comprehension
mutator returned nothing. The next mutant on that line, setting the
entire thing to None, was successful.
However, when running the mutation with the mutation id from cache, the `not in` mutation doesn't get applied to memory (because from cache the mutants are applied one at a time). This means the list comprehension mutant can proceed correctly. However, this throws off the indexing of the mutants, as the list comprehension mutant wasn't successful on the initial pass that populated the cache. As a result, the very last mutant (in this case, setting everything to None), is lost.