Converting loop mutation deepcopy into a shallow copy.
A single shallow copy doesn't work, because the actual loop children are
nested several layers deep inside the node object. What we really want
is a shallow copy of the container that holds the suite node's children,
and the child nodes themselves should not be copied. Added helper
function to do that copying, updated the one-loop and zero-while
mutations to use it (0-for doesn't need it since it operates on the for
statement and not the children). Also removed a unit test that did not
mutate loops.